Make Individual partial class
